requirements.txt
deepdiff
python-benedict
pyyaml